Formatted the code and added logs for better understanding
